# spring_ajax_jpa/openapi.yaml
openapi: 3.0.1
info:
  title: Game Management Service
  description: API for managing computer games, developers and publishers
  version: 1.0.0
servers:
  - url: http://localhost:8080/ComputerGames
...
paths:
...
  /developers:
    get:
      tags: [Developers]
      summary: Get all developers
      responses:
        '200':
          description: List of all developers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Developer'
                  
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
    post:
      tags: [Developers]
      summary: Create a new developer
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Developer name
                  example: CD Projekt Red
                country:
                  type: string
                  description: Developer country
                  example: Poland
                foundedDate:
                  type: string
                  format: date
                  description: Developer founded date
                  example: 1994-02-01
              required:
                - name
                - country
                - foundedDate
      responses:
        '201':
          description: Developer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
  /developers/{id}:
    delete:
      tags: [Developers]
      summary: Delete a developer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Developer deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '404':
          description: Developer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
    get:
      tags: [Developers]
      summary: Get a developer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Developer finded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Developer'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '403':
          description: Authentication granted but user does not have access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
                
        '404':
          description: Developer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationStatus'
...
components:
  schemas:
...
    Developer:
      type: object
      properties:
        id:
          type: integer
          description: Game developer identifier
          example: 1
        name:
          type: string
          description: Game developer name
          example: CD Projekt Red
        country:
          type: string
          description: Game developer country
          example: Poland
        foundedDate:
          type: string
          format: date
          description: Game developer founded data
          example: 1994-02-01
      required:
        - id
        - name
        - country
        - foundedDate
...
    OperationStatus:
      type: object
      properties:
        status:
          type: string
          description: Description of the operation execution status
          example: Operation error, invalid fields